feat: refactor styles add hooks
