+++ src/components/DoughnutChart/DoughnutChart.js
@@ -1,8 +1,8 @@
 import { Card, CardContent, CardHeader, Typography } from '@material-ui/core';
 import React from 'react';
 import { Doughnut } from 'react-chartjs-2';
+import useTransactions from '../../hooks/useTransactions';
 import useStyles from './styles';
-import useTransactions from './useTransactions';
 
 const DoughnutChart = ({ title }) => {
   const { total, chartData } = useTransactions(title);
