+++ src/styles.js
@@ -4,27 +4,16 @@ export default makeStyles(theme => ({
   root: {
     display: 'flex',
   },
-  main: {
-    background: theme.palette.background.paper,
-    [theme.breakpoints.up('sm')]: {
-      paddingBottom: '5%',
-    },
+  paper: {
+    padding: theme.spacing(2),
+    display: 'flex',
+    overflow: 'auto',
+    flexDirection: 'column',
   },
-  mobile: {
-    [theme.breakpoints.down('xs')]: {
-      margin: '5% 0',
-    },
-  },
-  last: {
+  fixedHeight: {
+    height: 250,
     [theme.breakpoints.down('sm')]: {
-      marginBottom: theme.spacing(3),
-      paddingBottom: '200px',
-    },
-  },
-  grid: {
-    paddingTop: '100px',
-    '& > *': {
-      margin: theme.spacing(2),
+      height: '100%',
     },
   },
 }));
